basic working flow working/ use json key value database to store sync files details

+++ cmd/delete.go
@@ -17,15 +17,15 @@ import (
 // deleteCmd represents the delete command
 var deleteCmd = &cobra.Command{
 	Use:   "rm",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Short: "store rm filenames....",
+	Long:  ``,
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println("delete called")
+
+		if len(args) == 0 {
+
+			return
+		}
 
 		deleteFile(args)
 
@@ -59,7 +59,7 @@ func deleteFile(files []string) {
 	body, _ := json.Marshal(FileBodyreq)
 
 	// Make the HTTP POST request
-	url := "http://" + URL + "apis/file/"
+	url := "http://" + URL + "/file/"
 	response, err := http.NewRequest("DELETE", url, bytes.NewBuffer(body))
 	if err != nil {
 		fmt.Println("Error making POST request:", err)
